requirements.txt
nltk
streamlit-keyup
st-annotated-text